requirements.txt
django~=3.1.7
djangorestframework-jwt~=1.11.0
djangorestframework~=3.12.2
PyJWT~=1.7.1